refactor: Move isMCPackage to RemapperBase

+++ src/main/java/space/libs/mixins/nbt/MixinCompressedStreamTools.java
@@ -64,7 +64,7 @@ public abstract class MixinCompressedStreamTools {
             CrashReport crashreport = CrashReport.makeCrashReport(exception, "Loading NBT data");
             CrashReportCategory crashreportcategory = crashreport.makeCategory("NBT Tag");
             crashreportcategory.addCrashSection("Tag name", "[UNNAMED TAG]");
-            crashreportcategory.addCrashSection("Tag type", Byte.valueOf(b0));
+            crashreportcategory.addCrashSection("Tag type", b0);
             throw new ReportedException(crashreport);
         }
     }
